workwave/requirement.txt
annotated-types==0.5.0
asgiref==3.7.2
Django==4.2.4
django-bootstrap-datepicker-plus==5.0.4
mysqlclient==2.2.0
pydantic==2.3.0
pydantic_core==2.6.3
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
